requirements.txt
PyYAML
rich

